requirements.txt
git+https://github.com/Rapptz/discord.py
aiohttp==3.7.4
python-dotenv
discord.py
yt-dlp
